Change approach: use core edits to check bumpability
Apply patches to `update_last_post()` and `update_thread_data()` in
`inc/functions.php` to have them check whether the thread is bumpable,
and if not, refrain from updating last post data.

This ensures that we catch situations in which these functions are
called that we were not previously catching, such as deletion of posts
in a thread.